fix: saved session logs config
